azfunction_inventory/requirements.txt
azure-core==1.36.0
azure-functions==1.24.0
azure-storage-blob==12.27.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
isodate==0.7.2
MarkupSafe==3.0.3
numpy==2.3.5
pandas==2.3.3
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.4
wheel==0.45.1
